Create Beacon
I have added the necessary comments for the Arduino file. Once the "beginRide()" function is executed for a RSU from receiving a message from the beacon unit, the beginRide variable switches to true. So in the main loop, "while(beginRide)" should run forever since it is a while(true) statement. However, for some reason, it exits that function and I cannot figure out why. Please help!