Work on README.md
This github repository will account for a large percentage of our handover to the next team, The README.md will serve as a means to help the next team get caught up to speed with the project. It will provide information that is project pertinent as well as any information that we accrued along the way that we wish we would have known earlier.